src/requirements.txt
async-timeout
attrs
azure-core
azure-functions
azure-functions-durable
azure-storage-blob
certifi
charset-normalizer
cryptography
frozenlist
furl
gitdb
GitPython
idna
isodate
orderedmultidict
pycparser
python-dateutil
requests
six
smmap
typing_extensions
urllib3